refactor: move create_collection() to create_stac.py
